FL: made utils folder with smoothstep func
